+++ iam-login-service/src/main/java/it/infn/mw/iam/persistence/repository/client/IamClientRepository.java
@@ -22,12 +22,12 @@ import java.util.Optional;
 import org.mitre.oauth2.model.ClientDetailsEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface IamClientRepository extends PagingAndSortingRepository<ClientDetailsEntity, Long>,
+public interface IamClientRepository extends JpaRepository<ClientDetailsEntity, Long>,
     JpaSpecificationExecutor<ClientDetailsEntity> {
 
   Optional<ClientDetailsEntity> findByClientId(String clientId);