+++ iam-persistence/src/main/java/it/infn/mw/iam/persistence/repository/IamOidcIdRepository.java
@@ -17,12 +17,12 @@ package it.infn.mw.iam.persistence.repository;
 
 import java.util.Optional;
 
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 
 import it.infn.mw.iam.persistence.model.IamOidcId;
 
-public interface IamOidcIdRepository extends CrudRepository<IamOidcId, Long> {
+public interface IamOidcIdRepository extends JpaRepository<IamOidcId, Long> {
 
   Optional<IamOidcId> findByIssuerAndSubject(@Param("issuer") String issuer,
       @Param("subject") String subject);