+++ iam-persistence/src/main/java/it/infn/mw/iam/persistence/repository/IamSshKeyRepository.java
@@ -17,12 +17,12 @@ package it.infn.mw.iam.persistence.repository;
 
 import java.util.Optional;
 
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 
 import it.infn.mw.iam.persistence.model.IamSshKey;
 
-public interface IamSshKeyRepository extends CrudRepository<IamSshKey, Long> {
+public interface IamSshKeyRepository extends JpaRepository<IamSshKey, Long> {
 
   Optional<IamSshKey> findByFingerprint(@Param("fingerprint") String fingerprint);